Changed Prime-Probe example structure
